Split term database into term util.
