+++ src/expr/node.cpp
@@ -22,7 +22,8 @@
 #include "base/output.h"
 #include "expr/attribute.h"
 
-#include "theory/quantifiers/term_database.h"
+// FIXME : can be removed if we properly handle #1205
+#include "theory/quantifiers/term_util.h"
 
 
 using namespace std;
@@ -114,10 +115,11 @@ bool NodeTemplate<ref_count>::hasBoundVar() {
         hasBv = (*i).hasBoundVar();
       }
       if( !hasBv ){
-        //FIXME : this is a hack to handle synthesis conjectures
+        //FIXME : this is a hack to handle synthesis conjectures (issue #1205)
         // the issue is that we represent second-order quantification in synthesis conjectures via a Node:
         //  exists x forall y P[f,y], where x is a dummy variable that maps to f through attribute SygusSynthFunVarListAttributeId
         //  when asked whether a node has a bound variable, we want to treat f as if it were a bound (second-order) variable. -AJR
+        // This can be solved once higher-order is merged by doing explicit higher-order quantification.
         if( getKind()==kind::APPLY_UF && getOperator().hasAttribute(theory::SygusSynthFunVarListAttribute()) ){
           hasBv = true;
         }
