+++ src/theory/quantifiers/rewrite_engine.cpp
@@ -24,6 +24,7 @@
 #include "theory/quantifiers/quant_conflict_find.h"
 #include "theory/quantifiers/quant_util.h"
 #include "theory/quantifiers/term_database.h"
+#include "theory/quantifiers/term_util.h"
 #include "theory/theory_engine.h"
 
 using namespace CVC4;
@@ -263,7 +264,7 @@ void RewriteEngine::registerQuantifier( Node f ) {
             body_c.push_back( d_rr[f][1][j].negate() );
           }
         }
-      }else if( d_rr[f][1]!=d_quantEngine->getTermDatabase()->d_true ){
+      }else if( d_rr[f][1]!=d_quantEngine->getTermUtil()->d_true ){
         if( MatchGen::isHandled( d_rr[f][1] ) ){
           body_c.push_back( d_rr[f][1].negate() );
         }
@@ -292,7 +293,7 @@ bool RewriteEngine::checkCompleteFor( Node q ) {
 Node RewriteEngine::getInstConstNode( Node n, Node q ) {
   std::map< Node, Node >::iterator it = d_inst_const_node[q].find( n );
   if( it==d_inst_const_node[q].end() ){
-    Node nn = d_quantEngine->getTermDatabase()->getInstConstantNode( n, q );
+    Node nn = d_quantEngine->getTermUtil()->getInstConstantNode( n, q );
     d_inst_const_node[q][n] = nn;
     return nn;
   }else{
